requirements.txt
openai==1.86.0
streamlit==1.45.1
pillow==11.2.1
requests==2.32.3
PyAudio==0.2.14
Wave==0.0.2